Reorganizes tests directory to fit the following structure: - tests/   - parse/   - scrape/
